requirement.txt
pyaudio
wave
flask
flask_socketio
azure-core
azure-cognitiveservices-speech